chore(hybrid): rolldown isolation
